Incremental patch of Rabin Karp /* There was a bug in the last patch that was generating hash for the remaining contents in the previous buffer, considering it as new chunk even it does not match the finger print.
Vector is implemented to overcome from this problem, by storing
the remaining contents of all the previous buffer which did not
match the finger print as well as the buffer content which matched
the finger print. Then it takes the vector which contains the actual
chunk and calculate the hash and updates the stores.

Looking into the compatibility issue with different file types and
design of get_variable_chunk function.
*/

Signed-off-by: Sukumar <sukumarpoojari92@gmail.com>